Add tabs in WorkflowInvocationState
Also refactor tab contents into individual components
