+++ client/src/components/WorkflowInvocationState/WorkflowInvocationDetails.vue
@@ -1,14 +1,10 @@
 <template>
     <div v-if="invocation">
-        <div v-if="Object.keys(invocation.input_step_parameters).length">
-            <details class="invocation-parameters-details">
-                <summary><b>Parameters</b></summary>
+        <b-tabs lazy>
+            <b-tab v-if="Object.keys(invocation.input_step_parameters).length" title="Parameters">
                 <parameter-step :parameters="Object.values(invocation.input_step_parameters)" />
-            </details>
-        </div>
-        <div v-if="Object.keys(invocation.inputs).length">
-            <details class="invocation-inputs-details">
-                <summary><b>Inputs</b></summary>
+            </b-tab>
+            <b-tab v-if="Object.keys(invocation.inputs).length" title="Inputs">
                 <div
                     v-for="(input, key) in invocation.inputs"
                     :key="input.id"
@@ -16,29 +12,20 @@
                     <b>{{ dataInputStepLabel(key, input) }}</b>
                     <generic-history-item :item-id="input.id" :item-src="input.src" />
                 </div>
-            </details>
-        </div>
-        <div v-if="Object.keys(invocation.outputs).length">
-            <details class="invocation-outputs-details">
-                <summary><b>Outputs</b></summary>
+            </b-tab>
+            <b-tab v-if="Object.keys(invocation.outputs).length" title="Outputs">
                 <div v-for="(output, key) in invocation.outputs" :key="output.id">
                     <b>{{ key }}:</b>
                     <generic-history-item :item-id="output.id" :item-src="output.src" />
                 </div>
-            </details>
-        </div>
-        <div v-if="Object.keys(invocation.output_collections).length">
-            <details class="invocation-output-collections-details">
-                <summary><b>Output Collections</b></summary>
+            </b-tab>
+            <b-tab v-if="Object.keys(invocation.output_collections).length" title="Output Collections">
                 <div v-for="(output, key) in invocation.output_collections" :key="output.id">
                     <b>{{ key }}:</b>
                     <generic-history-item :item-id="output.id" :item-src="output.src" />
                 </div>
-            </details>
-        </div>
-        <div v-if="workflow">
-            <details v-if="workflow" class="invocation-steps-details">
-                <summary><b>Steps</b></summary>
+            </b-tab>
+            <b-tab v-if="workflow" title="Steps">
                 <workflow-invocation-step
                     v-for="step in Object.values(workflow.steps)"
                     :key="step.id"
@@ -46,8 +33,8 @@
                     :ordered-steps="orderedSteps"
                     :workflow="workflow"
                     :workflow-step="step" />
-            </details>
-        </div>
+            </b-tab>
+        </b-tabs>
     </div>
 </template>
 <script>
@@ -66,11 +53,8 @@ export default {
     },
     props: {
         invocation: {
+            type: Object,
             required: true,
-        },
-        invocationAndJobTerminal: {
-            required: true,
-            type: Boolean,
         },
     },
     computed: {
